+++ bokeh/_templates/plot_js.js
@@ -9,3 +9,4 @@ var all_models = {{ all_models }};
 Bokeh.load_models(all_models);
 var model = Bokeh.Collections(modeltype).get(modelid);
 var view = new model.default_view({model: model, el: '#{{ elementid }}'});
+Bokeh.index[modelid] = view