+++ bokeh/mixins.py
@@ -4,12 +4,12 @@ from .properties import HasProps, ColorSpec, DataSpec, Enum, DashPattern, Int, S
 from .enums import LineJoin, LineCap, FontStyle, TextAlign, TextBaseline
 
 class FillProps(HasProps):
-    """ Mirrors the BokehJS properties.fill_properties class """
+    """ Mirrors the BokehJS properties.Fill class. """
     fill_color = ColorSpec("gray")
     fill_alpha = DataSpec(1.0)
 
 class LineProps(HasProps):
-    """ Mirrors the BokehJS properties.line_properties class """
+    """ Mirrors the BokehJS properties.Line class. """
     line_color = ColorSpec("black")
     line_width = DataSpec
     line_alpha = DataSpec(1.0)
@@ -19,11 +19,11 @@ class LineProps(HasProps):
     line_dash_offset = Int(0)
 
 class TextProps(HasProps):
-    """ Mirrors the BokehJS properties.text_properties class """
+    """ Mirrors the BokehJS properties.Text class. """
     text_font = String("Helvetica")
-    text_font_size = String("10pt")
+    text_font_size = String("12pt")
     text_font_style = Enum(FontStyle)
-    text_color = ColorSpec("black")
+    text_color = ColorSpec("#444444")
     text_alpha = DataSpec(1.0)
     text_align = Enum(TextAlign)
-    text_baseline = Enum(TextBaseline)
+    text_baseline = Enum(TextBaseline, default="bottom")