+++ examples/plotting/file/scatter_selection.py
@@ -1,6 +1,6 @@
 import numpy as np
 from bokeh.plotting import *
-import bokeh
+from bokeh.objects import BoxSelectTool
 N = 100
 
 x = np.linspace(0, 4*np.pi, N)
@@ -10,26 +10,26 @@ output_file("scatter_selection.html", title="scatter_selection.py example")
 
 figure(tools="pan,wheel_zoom,box_zoom,reset,previewsave,select")
 
-l = scatter(x,y, 
+l = scatter(x,y,
     color="red",
     title="select on mouseup"
 )
-select_tool = [t for t in l.tools if isinstance(t, bokeh.objects.BoxSelectTool)][0]
+select_tool = curplot().select(dict(type=BoxSelectTool))
 select_tool.select_every_mousemove = False
 
-l = scatter(x,y, 
-    marker="square", color="green", 
+l = scatter(x,y,
+    marker="square", color="green",
     title='select on mousemove'
     )
-select_tool = [t for t in l.tools if isinstance(t, bokeh.objects.BoxSelectTool)][0]
+select_tool = curplot().select(dict(type=BoxSelectTool))
 
-scatter(x,y, 
+scatter(x,y,
     color="#FF00FF", nonselection_fill_color="#FFFF00", nonselection_fill_alpha=1
 )
 
 select_tool.select_every_mousemove = True
 
-scatter(x,y, 
+scatter(x,y,
     marker="square", color="blue", name="scatter_selection_example"
 )